+++ __mocks__/axios.js
@@ -1,0 +1,7 @@
+export default {
+    create: jest.fn(() => {
+        return {
+            post: jest.fn(() => Promise.resolve({data: {}}))
+        }
+    })
+}